chore: refactoring
LiveDocHelper renamed to DocumentDataHelper
DocumentData moved to external class
